Factorial.java
Factorial Calculation Program

This Java program calculates the factorial of a given number `n`. The factorial of a non-negative integer `n` (denoted as `n!`) is the product of all positive integers less than or equal to `n`. For example, `5! = 5 * 4 * 3 * 2 * 1 = 120`.

Program Functionality:

1. Prompts the user to enter the value of `n`.
2. Calculates the factorial of `n` using a loop.
3. Prints the factorial value.

#### Example Output:

```
enter the value of n
5
fact=120
```

If the value of `n` is 0, the program prints `fact=0`.

This program efficiently computes the factorial of a given number using a simple loop structure.